perf: eliminate nav delay on About and Compare pages
- About page.tsx: convert from async Server Component to sync, extract
  all content to new AboutContent client component using useTranslations
- Compare page.tsx: remove unused getTranslations call (translations are
  already consumed by ComparePageHeader/CompareTable client components)
- Add compare/loading.tsx skeleton to give instant visual feedback while
  searchParams resolves on the server
